rename locale file
