+++ keyboards/chonkerkeys/core.c
@@ -2,7 +2,7 @@
 #include "keyconfig.h"
 #include "layer_type.h"
 #include "rgb_strands/rgb_strands.h"
-#include "default_locales.h"
+#include "chonkerkeys_locales.h"
 #include <math.h>
 
 #define KEY_MACROS_MAX_COUNT  3
